pinyin-immersion-app/requirements.txt
streamlit
python-dotenv
google-genai
gTTS
pandas